This is an updated version of #3457 by @slaren 

Also, adds support for quantum K cache: #4312 

It's adapted to the latest code on `master`. Should work with all model arches

- `-ngl` now goes up to `n_layer + 1` instead of `n_layer + 3`
- use `--no-kv-offload` or `-nkvo` to disable KV cache offloading to GPU

## `llama.h` API Changes

- `struct llama_context_params`
  - add `enum ggml_type type_k; // data type for K cache`
  - add `enum ggml_type type_v; // data type for V cache`
  - add `bool offload_kqv;      // whether to offload the KQV ops (including the KV cache) to GPU`
  - remove `bool f16_kv`
- `LLAMA_SESSION_VERSION` bump to 3

## Limitations

- Context shift does not work with quantum K cache
- V cache quantization is not yet supported

TODO:
- [x] add command-line args to disable KV offloading (enabled by default)
- [ ] extensive testing with different models and different `-ngl` values
- [ ] double-check that the KV shift on context overflow works correctly (initial tests indicate that it does)
- [x] update state save/load to the new KV cache and bump `LLAMA_SESSION_VERSION`