@oobabooga 2023-12-03
Here is a test on my hardware (RTX 3090, 3200 MHz RAM).

Command:

```
./main \
  -m ../models/wizardlm-70b-v1.0.Q4_K_S.gguf \
  -e \
  -t 6 \
  -tb 12 \
  -n 256 \
  -c 4096 \
  -ngl 42 \
  --ignore-eos \
  -p "Because of the rapid pace of improvement of large language models, evaluation benchmarks have suffered from short lifespans, with state of the art models quickly saturating existing benchmarks, exceeding the performance of human annotators, leading to efforts to replace or augment the benchmark with more challenging tasks. In addition, there are cases of shortcut learning wherein AIs sometimes cheat on multiplechoice tests by using statistical correlations in superficial test question wording in order to guess the correct responses, without necessarily understanding the actual question being asked.\n\nSome datasets have been constructed adversarially, focusing on particular problems on which extant language models seem to have unusually poor performance compared to humans. One example is the TruthfulQA dataset, a question answering dataset consisting of  questions which language models are susceptible to answering incorrectly by mimicking falsehoods to which they were repeatedly exposed during training. For example, an LLM may answer No to the question Can you teach an old dog new tricks because of its exposure to the English idiom you cant teach an old dog new tricks, even though this is not literally true.\n\nAnother example of an adversarial evaluation dataset is Swag and its successor, HellaSwag, collections of problems in which one of multiple options must be selected to complete a text passage. The incorrect completions were generated by sampling from a language model and filtering with a set of classifiers. The resulting problems are trivial for humans but at the time the datasets were created state of the art language models had poor accuracy on them. For example:"
```

## PR (c80b8a2bff73239b88700fd0b8d32860ad1c6986)

Peak VRAM: 23049MiB

```
llama_print_timings:        load time =   30315,12 ms
llama_print_timings:      sample time =      66,38 ms /   256 runs   (    0,26 ms per token,  3856,82 tokens per second)
llama_print_timings: prompt eval time =    3365,29 ms /   332 tokens (   10,14 ms per token,    98,65 tokens per second)
llama_print_timings:        eval time =  118578,99 ms /   255 runs   (  465,02 ms per token,     2,15 tokens per second)
llama_print_timings:       total time =  122225,02 ms
```

## Master

Peak VRAM: 22377MiB

```
llama_print_timings:        load time =   31612,69 ms
llama_print_timings:      sample time =      71,63 ms /   256 runs   (    0,28 ms per token,  3573,87 tokens per second)
llama_print_timings: prompt eval time =    4140,64 ms /   332 tokens (   12,47 ms per token,    80,18 tokens per second)
llama_print_timings:        eval time =  119670,77 ms /   255 runs   (  469,30 ms per token,     2,13 tokens per second)
llama_print_timings:       total time =  124069,09 ms
```

-----

The output is stable and coherent in both cases. The conclusion is that the changes in this PR increase the tokens/second for prompt evaluation by +23% while increasing the VRAM usage by +672MB in this particular setup.