@kalomaze 2023-12-03
Both thread counts seem to matter, perhaps equally now. In the past it was said that you should focus on batched threads instead of regular threads... But if I set both to 7, I get better speeds. (My CPU has 4 cores and 8 threads)

Threads count 7, Batched threads count 7: 127 t/s prompt processing, 2.5 t/s generation speed
Thread count 2, Batched threads count 7: 137 t/s prompt processing, 1.3t/s generation speed