@kalomaze 2023-12-07
> Plan to merge this tomorrow - haven't tested much yet. Hope it does not break something
> 
> The CUDA performance with quantum cache is a bit disappointing. Hopefully we will fix this in the future

I have been testing my sampler experiments using it and I don't notice any breaking changes.