@kalomaze 2023-12-07
> The CUDA performance with quantum cache is a bit disappointing.

Thankfully, I'd argue "disappointing" is potentially misleading, looking at the benchmark provided in the other PR. It was my presumption that quantization actually reduces throughput because of the overhead of reducing precision, but if it means you can fit a model entirely into VRAM that wouldn't fit before, or _more layers_ of a model into VRAM specifically, _that alone_ is much more important than the actual compute being slower because VRAM is leaps and bounds faster than regular RAM.