@ggerganov 2023-12-07
> I have been testing my sampler experiments using it and I don't notice any breaking changes, but that was before you added the K layer quantization stuff apparently if I read the other PR right?

Yes, thank you for testing. The changes that I expect to break something are related to the "per-layer KV cache". The other changes from #4312 should be good I think.

"breaking change" means that the API has changed - i.e. it is not backwards compatible.

> Thankfully, I'd argue "disappointing" is potentially misleading, looking at the benchmark provided in the other PR.

It's disappointing because it could be much better, but the quantum kernels are not well adapted for multi-head tensors (i.e. `ne[2] > 1`). In that case, we are doing unnecessary copies and launching separate kernels for each head, while it could be done more efficiently in a single pass. I tried to make a quick workaround for the single GPU case, but it's ugly.