@x4080 2023-12-07
Is this benefits apple silicon as well ?