@CyborgArmy83 2023-12-07
> Is this benefits apple silicon as well ?

Curious about this as well. Would be nice if it could speed up prompt processing on Apple Silicon.