@Ph0rk0z 2023-12-08
So for full offload this will not make much difference? I think I have to test this via the built in server rather than python bindings. I would hope that 8bit KV would have helped something.