@Dampfinchen 2023-12-10
@LostRuins That's because you are just processing 200 tokens, while I am using a full 2K prompt and also generate 180 Tokens instead of 31. The difference offloaded KV cache makes, is huge, both in generation and prompt processing speed.

22 Layers, per-layer KV cache offloaded, 6 GB 2060

![22layersperlayerkvcache](https://github.com/ggerganov/llama.cpp/assets/59751859/f64c5110-36be-4a27-831b-ce9ce4de8b29)

25 Layers, per layer KV Cache disabled

![25layers,nokvoffload](https://github.com/ggerganov/llama.cpp/assets/59751859/88ed51ef-9b40-4c2f-b674-1c56c314dda4)