@Ph0rk0z 2023-12-14
Whelp. I forced on kv_cache offloading in the llama python bindings. It says it's using type f16 as before. Prompt processing speeds on multi-gpu 70b are 1/4 of previous speed and now I get max 10 t/s. My 3090 turned into a P40. 

Unsure if it's really offloaded. Or if it's something else. Could be why I'm getting terrible mistral speeds as well since it spans GPU.